Space object added
The space object wil act as the world in which all logic is handled. The asiPrototype file is just an renderer. The goal is to write the Space class, and all it's subclasses, so that it is easily portable to the Arduino. The Space class and it subclasses only contains the game logic, no visualisation what so ever